job-exec: move shell-exit handling to jobinfo
Problem: the leader-shell-exit handling (post the shell-exit event and
arm a timer to catch stragglers) is contained in the bulk-exec
implementation, so it cannot be reused by another exec implementation.

Move it into jobinfo_post_shell_exit() in job-exec.c, alongside the
sibling timers it resembles.
